New Graphml and Fixed a bug
Added a new graphml to test (as it clearly should form two distinct communities). Also fixed a bug in the Label Propagation Algorithm where every time a node went through label propagation, it's label would simply change to the label of its last neighbor node (instead of changing the label to the label that the majority of its neighbor nodes have).